backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
djoser==2.2.0
social-auth-app-django==5.4.0
python-jose==3.3.0